pyproject.toml
[project]
name = "gemini-apps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.3.4",
    "faiss-cpu>=1.13.0",
    "google-genai>=1.50.1",
    "groq>=0.34.1",
    "langchain>=1.0.7",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.0.3",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "pdf2image>=1.17.0",
    "pypdf>=6.3.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "youtube-transcript-api>=1.2.3",
]
